Split i18n dictionaries
