+++ frontend/lib/i18n/dictionary-data/en.ts
@@ -1,0 +1,26 @@
+import type { Dictionary } from '../dictionary-types';
+import { footer, nav } from './en-layout';
+import { home } from './en-home';
+import { calculator, pricing } from './en-pricing';
+import { models, workflows } from './en-workflows-models';
+import { about, blog, contact, docs, examples } from './en-content';
+import { changelog, legal, status, systemMessages } from './en-legal-status';
+
+export const en: Dictionary = {
+  nav,
+  footer,
+  home,
+  pricing,
+  calculator,
+  workflows,
+  models,
+  examples,
+  docs,
+  blog,
+  about,
+  contact,
+  legal,
+  changelog,
+  status,
+  systemMessages,
+};
