+++ frontend/lib/i18n/dictionary-data/fr.ts
@@ -1,0 +1,4 @@
+import { en } from './en';
+import type { Dictionary } from '../dictionary-types';
+
+export const fr: Dictionary = en;
